[FEATURE-12] Refactor Host project
