+++ src/Host/Pages/Error.cshtml.cs
@@ -2,7 +2,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using System.Diagnostics;
 
-namespace YAGO.WebProjectTemplate.Web.Pages
+namespace YAGO.WebProjectTemplate.Host.Pages
 {
 	[ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
 	public class ErrorModel : PageModel
