+++ src/Host/Pages/_ViewImports.cshtml
@@ -1,3 +1,3 @@
-﻿@using YAGO.WebProjectTemplate.Web
-@namespace YAGO.WebProjectTemplate.Web.Pages
+﻿@using YAGO.WebProjectTemplate.Host
+@namespace YAGO.WebProjectTemplate.Host.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
